agents/stock-analysis-extended/requirements.txt
# Core dependencies
crewai>=0.1.0
crewai-tools>=0.1.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-openai>=0.0.5
langgraph>=0.0.1

# Financial data APIs
yfinance>=0.2.28
alpha-vantage>=2.3.1
pandas-datareader>=0.10.0
finnhub-python>=2.4.0
polygon-api-client>=1.13.0

# Technical analysis
pandas>=2.0.0
numpy>=1.24.0
ta-lib>=0.4.28
pandas-ta>=0.3.14b0
plotly>=5.17.0
matplotlib>=3.7.0

# Sentiment analysis
newsapi-python>=0.2.7
tweepy>=4.14.0
praw>=7.7.1
textblob>=0.17.1
vaderSentiment>=3.3.2
transformers>=4.35.0

# Data processing
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.8.0
asyncio>=3.4.3

# Database and caching
redis>=5.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.5.0

# Message queue
pika>=1.3.0
celery>=5.3.0
kombu>=5.3.0

# Monitoring and logging
prometheus-client>=0.18.0
structlog>=23.2.0
sentry-sdk>=1.38.0

# Web framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Environment and config
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Utilities
schedule>=1.2.0
retry>=0.9.2
tenacity>=8.2.0
pytz>=2023.3